Add support for importing dumps after downloading them
This only works for full dumps as of now.

+++ listenbrainz_spark/data/import_data/import_dump.py
@@ -33,7 +33,7 @@ def _is_json_file(filename):
     return filename.endswith('.json')
 
 
-def _process_json_file(filename, data_dir, hdfs_path):
+def _process_json_file(filename, data_dir, hdfs_path, full=True):
     """ Process a file containing listens from the ListenBrainz dump and add listens to
     appropriate dataframes.
     """
@@ -49,11 +49,14 @@ def _process_json_file(filename, data_dir, hdfs_path):
         dest_path = os.path.join(data_dir, year, '{}.parquet'.format(str(month)))
 
     print("Uploading to %s..." % dest_path)
-    file_df.write.format('parquet').save(config.HDFS_CLUSTER_URI + dest_path)
+    if full:
+        file_df.write.format('parquet').save(config.HDFS_CLUSTER_URI + dest_path)
+    else:
+        file_df.write.mode('append').parquet(config.HDFS_CLUSTER_URI + dest_path)
     print("File processed in %.2f seconds!" % (time.time() - start_time))
 
 
-def copy_to_hdfs(archive, threads=8):
+def copy_to_hdfs(archive, full=True, threads=8):
 
     """ Create Spark Dataframes from a listens dump and save it to HDFS.
 
@@ -65,10 +68,20 @@ def copy_to_hdfs(archive, threads=8):
     pxz_command = ['pxz', '--decompress', '--stdout', archive, '-T{}'.format(threads)]
     pxz = subprocess.Popen(pxz_command, stdout=subprocess.PIPE)
     destination_path = os.path.join('/', 'data', 'listenbrainz')
-    if FORCE:
+    if full and FORCE:
         print('Removing data directory if present...')
         hdfs_connection.client.delete(destination_path, recursive=True)
         print('Done!')
+
+    dump_id = int(os.path.split(archive)[1].split('-')[2])
+    if not full:
+        hdfs_connection.client.download(os.path.join(destination_path, 'DATA_VERSION'), tmp_dump_dir)
+        with open(os.path.join(tmp_dump_dir, 'DATA_VERSION')) as f:
+            prev_dump_id = int(f.read().strip())
+
+        if dump_id != prev_dump_id + 1:
+            print("Incorrect incremental dump being imported, expected %d, got %d, exiting..." % prev_dump_id + 1, dump_id)
+            raise SystemExit("Incorrect incremental dump being imported")
 
     file_count = 0
     total_time = 0.0
@@ -80,7 +93,7 @@ def copy_to_hdfs(archive, threads=8):
                 tar.extract(member)
                 tmp_hdfs_path = os.path.join(tmp_dump_dir, member.name)
                 hdfs_connection.client.upload(hdfs_path=tmp_hdfs_path, local_path=member.name)
-                _process_json_file(member.name, destination_path, tmp_hdfs_path)
+                _process_json_file(member.name, destination_path, tmp_hdfs_path, full=full)
                 hdfs_connection.client.delete(tmp_hdfs_path)
                 os.remove(member.name)
                 file_count += 1
@@ -89,13 +102,20 @@ def copy_to_hdfs(archive, threads=8):
                 total_time += time_taken
                 average_time = total_time / file_count
                 print("Total time: %.2f, average time: %.2f" % (total_time, average_time))
+
+    with open(os.path.join(tmp_dump_dir, 'DATA_VERSION'), 'w') as f:
+        f.write(str(dump_id) + "\n")
+    hdfs_connection.client.upload(
+        hdfs_path=os.path.join(destination_path, 'DATA_VERSION'),
+        local_path=os.path.join(tmp_dump_dir, 'DATA_VERSION'),
+        overwrite=True,
+    )
+
     hdfs_connection.client.delete(tmp_dump_dir, recursive=True)
     shutil.rmtree(tmp_dump_dir)
 
 
-def main(app_name, archive):
-    listenbrainz_spark.init_spark_session(app_name)
-    hdfs_connection.init_hdfs(config.HDFS_HTTP_URI)
+def main(archive, full=True):
     print('Copying extracted dump to HDFS...')
     copy_to_hdfs(archive)
     print('Done!')
